test(unit): gameStore を分割しスナップショットテストを追加する
gameStore.spec.ts が 260 行・29 テストで分割ルールに抵触していたため
責務単位で 5 ファイルに分割する。
あわせて VCell・VGameScore・VMain のスナップショットテストを追加し
意図しない HTML 変更を検知できるようにする。
